requirements.txt
requests
dj-database-url
python-dotenv
django-ses